refactor: change CommonAbility protected to private

+++ src/main/java/ability/domain/Ability.java
@@ -77,7 +77,7 @@ public class Ability {
         return status.currentStatus();
     }
 
-    public int currentAttack() {
+    public double currentAttack() {
         return attack.currentAttack();
     }
